Update code according to v3 changes
